+++ pantheon/src/test/java/tech/pegasys/pantheon/cli/options/SynchronizerOptionsTest.java
@@ -34,9 +34,7 @@ public class SynchronizerOptionsTest
     return SynchronizerConfiguration.builder()
         .fastSyncPivotDistance(SynchronizerConfiguration.DEFAULT_PIVOT_DISTANCE_FROM_HEAD + 10)
         .fastSyncFullValidationRate(SynchronizerConfiguration.DEFAULT_FULL_VALIDATION_RATE / 2)
-        // Min peers are currently handled outside of SynchronizerOptions
-        //      .fastSyncMinimumPeerCount(SynchronizerConfiguration.DEFAULT_FAST_SYNC_MINIMUM_PEERS
-        // + 2)
+        .fastSyncMinimumPeerCount(SynchronizerConfiguration.DEFAULT_FAST_SYNC_MINIMUM_PEERS + 2)
         .worldStateHashCountPerRequest(
             SynchronizerConfiguration.DEFAULT_WORLD_STATE_HASH_COUNT_PER_REQUEST + 2)
         .worldStateRequestParallelism(
@@ -84,5 +82,10 @@ public class SynchronizerOptionsTest
   SynchronizerConfiguration getDomainObjectFromPantheonCommand() {
     verify(mockControllerBuilder).synchronizerConfiguration(syncConfigurationCaptor.capture());
     return syncConfigurationCaptor.getValue();
+  }
+
+  @Override
+  protected List<String> getFieldsToIgnore() {
+    return Arrays.asList("fastSyncMinimumPeerCount");
   }
 }
